New tests for signMultiSigUTXO (previously SignMultiSigUTXO)
signMultiSigUTXO now also calls validateSigScript() to ensure the sigscript
it generates will execute later on.

New SignTx() function which signs all inputs of a given tx.

Plus removal of createTxWithInputAmounts, which is no longer needed as we
have the more flexible createDecoratedTx.

And last but not least, some fixes for issues pointed out by go vet.